% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xls_reading.R
\name{add_focus_name}
\alias{add_focus_name}
\title{Adds Column with Name of Condition of Health Focus}
\usage{
add_focus_name(df)
}
\arguments{
\item{df}{a dataframe}
}
\description{
TODO: perhaps there can be a generalization to handle add_focus_name,
 add_state_names
}

