% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xls_reading.R
\name{add_states}
\alias{add_states}
\title{Identify Rows Beginning State-related Section}
\usage{
add_states(df)
}
\arguments{
\item{df}{a dataframe}
}
\description{
Assumes there is a column named "Objective"
}

